+++ libs/platform/src/lib/table/components/table-view-settings-dialog/filtering/filter-single-select.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output, ViewEncapsulation } from '@angular/core';
 
 import { CollectionFilter } from '../../../interfaces/collection-filter.interface';
 import { TableFilterSelectOption } from '../../../interfaces/selection-value.interface';
@@ -12,7 +12,8 @@ import { NOT_FILTERED_OPTION_VALUE } from './constants';
 @Component({
     selector: 'fdp-filter-single-select',
     templateUrl: './filter-single-select.component.html',
-    changeDetection: ChangeDetectionStrategy.OnPush
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    encapsulation: ViewEncapsulation.None
 })
 export class FilterSingleSelectComponent {
     /** Selectable filter options */