+++ libs/platform/src/lib/table/models/tree-table.model.ts
@@ -1,0 +1,9 @@
+export type TreeLike<T> = T & {
+    _children?: TreeLike<T>[];
+};
+
+export interface GroupTableRowValueType {
+    field: string;
+    value: unknown;
+    count: number;
+}